requirements.txt
# Core Dependencies
python-dotenv>=0.19.0

# Video Processing Libraries
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Video Encoding/Decoding
av==11.0.0

# Database
psycopg2-binary>=2.9.0

# GUI Framework
PyQt5>=5.15.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
pylint==3.0.2
mypy==1.7.1
